requirements.txt
dlib==19.20.0
imutils==0.5.3
numpy==1.19.0
opencv-python==4.3.0.36
pika==1.3.1
Pillow==9.3.0
scipy==1.5.1
